refactor admin dashboard server handling
