+++ app/api/admin/orders/route.ts
@@ -1,0 +1,21 @@
+import { NextResponse } from 'next/server'
+import { createRouteHandlerClient } from '@supabase/auth-helpers-nextjs'
+import { createClient } from '@supabase/supabase-js'
+import { cookies } from 'next/headers'
+
+export async function GET() {
+  const supabase = createRouteHandlerClient({ cookies })
+  const { data: { user } } = await supabase.auth.getUser()
+  if (!user) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  const { data: profile } = await supabase.from('user_profiles').select('is_admin').eq('user_id', user.id).single()
+  if (!profile?.is_admin) return NextResponse.json({ error: 'Forbidden' }, { status: 403 })
+
+  const admin = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.SUPABASE_SERVICE_ROLE_KEY!)
+  const { data } = await admin
+    .from('orders')
+    .select('id, customer_email, amount, status, created_at, products (name)')
+    .order('created_at', { ascending: false })
+    .limit(50)
+
+  return NextResponse.json({ orders: data || [] })
+}
